Changing directions here
Was originally try to take a number, N, and determinte which (if any)
figurate number it could be. Turns out that is less than advantageous way
to do it, as factoring each number by each figurate formula is non trivial.
I should, instead, calculate all 4 digit numbers for each formula type. Also,
it may be possible for a number to multiple figurate types, such as both
and triangular, but I don't know that for sure right now.